pyproject.toml
[project]
name = "imlight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "glfw>=2.10.0",
    "jedi>=0.19.2",
    "moderngl>=5.12.0",
    "numpy>=2.3.4",
    "numpy-stl>=3.2.0",
    "ola",
    "pillow>=12.0.0",
    "pyopengl>=3.1.10",
    "pyrr>=0.10.3",
    "requests>=2.32.5",
    "imgui-bundle>=0.10.0",
]

[tool.uv]
package = true

[tool.uv.sources]
ola = { git = "https://github.com/N3rdL0rd/ola-py" }

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pyright>=1.1.407",
    "ruff>=0.14.3",
]

[project.scripts]
imlight = "imlight.main:main"

[[tool.mypy.overrides]]
module = ["untyped_package.*"]
follow_untyped_imports = true
